# DigitalLibrary/DigitalLibrary/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DigitalLibraryAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Publication/{id}": {
      "get": {
        "tags": [
          "Publication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Publication"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publication"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publication"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Publication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Publication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/PublicationType": {
      "get": {
        "tags": [
          "PublicationType"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PublicationType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              }
            }
          }
        }
      }
    },
    "/api/PublicationType/{id}": {
      "get": {
        "tags": [
          "PublicationType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PublicationType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "PublicationType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "publicationAuthors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationAuthor"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Borrower": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Category": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Format": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Lector": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Publication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "boolean"
          },
          "publicationYear": {
            "type": "integer",
            "format": "int32"
          },
          "publicationAuthors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationAuthor"
            },
            "nullable": true
          },
          "idCategory": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "idLector": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lector": {
            "$ref": "#/components/schemas/Lector"
          },
          "idPublishingHouse": {
            "type": "integer",
            "format": "int32"
          },
          "publishingHouse": {
            "$ref": "#/components/schemas/PublishingHouse"
          },
          "idPublicationType": {
            "type": "integer",
            "format": "int32"
          },
          "publicationType": {
            "$ref": "#/components/schemas/PublicationType"
          },
          "idFormat": {
            "type": "integer",
            "format": "int32"
          },
          "format": {
            "$ref": "#/components/schemas/Format"
          },
          "idBorrower": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "borrower": {
            "$ref": "#/components/schemas/Borrower"
          }
        },
        "additionalProperties": false
      },
      "PublicationAuthor": {
        "type": "object",
        "properties": {
          "idPublication": {
            "type": "integer",
            "format": "int32"
          },
          "publication": {
            "$ref": "#/components/schemas/Publication"
          },
          "idAuthor": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          }
        },
        "additionalProperties": false
      },
      "PublicationType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublishingHouse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}